Add exclude for blockstorage during resize (#211)
This patch skips the filesystem expansion when the volume was requested
in blockstorage mode, preventing issues with unexpected filesystems.

This patch resolved the problem that even in blockstorage mode, the
hcloud-csi tires to run a filesystem expansion when extending the cloud
volume. This can fail for LUKS volumes or volumes using another
unsupported filesystem and prevent the PVC from ever be successfully
expanded.

The easiest way to prevent that is to not fiddling with the filesystem
when the volume is used in block device mode.