Rework RBAC to only apply to CSI Controller
The ServiceAccount and associated ClusterRole + binding are only
necessary for the Controller StatefulSet. The DaemonSet does not require
any permissions to operate on objects in the cluster, since it is
merely responsible for registering the driver with kubelet.

https://github.com/kubernetes-csi/node-driver-registrar#required-permissions

Slightly arbitrary, but also rename the SA + role + binding to reflect
that it only applies to the controller.