Switch Controller to Deployment, plus other tweaks
The controller does not need to be a StatefulSet, it is not inherently
stateful. The various sidecars all perform leader election using the
standard k8s Lease approach.

As part of switching to Deployment, drop the governing serviceName field
which was referencing a non-existent Service anyway.

Also remove the node affinity preventing the Deployment from scheduling
replicas on Hetzner root servers. This is not necessary - the controller
should run on root servers just fine, since it's really just responsible
for issuing requests to Hetzner Cloud API in response to CSI Driver
requests. It's only the Node DaemonSet that should not be scheduled on
root servers.

Finally, drop the privileged security context - the controller is not
performing any privileged operations.