Remove HCLOUD_TOKEN from node DaemonSet
The final usage of hcloud API required on the node side was to determine
the location of the server to provide a topology hint to the controller
when creating a new volume.

Instead, the node now relies on the metadata service to determine the ID
and location of the server. I understand there has been issues with the
usage of the metadata service in the past, but from combing through the
old tickets, I'm reasonably confident that stemmed from not having a
distinction between the node and controller parts, which we now have. As
I pointed out in the commit to split the node and controller into
separate binaries, we can always assume the node part is running on a
hcloud server. That's kinda the entire point of it, really. So we should
be able to rely on the metadata service being reachable.

So now the node is no longer making any hcloud API calls whatsoever, and
the HCLOUD_TOKEN and KUBE_NODE_NAME env vars have been removed entirely.
Architecturally, the node is now in the state I was hoping to get it to:
it's a dumb, unprivileged drone that does the minimum on-node actions,
and nothing more.