Rework Node topology and remove HCLOUD_TOKEN use
The final usage of hcloud API required on the node side was to determine
the location of the server to provide a topology hint to the controller
when creating a new volume.

It would be nice if the hcloud metadata service had an endpoint to get
the DC location of the server but alas it does not exist, yet. Hmmmm...

Instead, the node now relies on the metadata service to determine the
ID of the server it's running on. I know in the past there's been issues
with the usage of the metadata service, but from combing through the old
tickets, I'm reasonably confident that stemmed from not having a
distinction between the node and controller parts, which we now have. As
I pointed out in the commit to split the node and controller into
separate binaries, we can always assume the node part is running on a
hcloud server (that's kinda the entire point of it, really), and so we
can rely on the metadata service being reliably reachable.

So with the server ID in hand, the node now provides this in NodeGetInfo
which is then provided to the ControllerCreateVolume call. At this point
the controller uses that server ID hint (if it exists) to determine the
correct location to create the volume in. The topology stuff was already
implemented as a "best effort" deal anyways, with a fallback to use the
location of the controller. So I continue that trend - if the server ID
is present but can't be looked up for some reason, do the fallback.

So now the node is no longer making any hcloud API calls whatsoever, and
the HCLOUD_TOKEN and KUBE_NODE_NAME env vars have been removed entirely.
Architecturally, the node is now in the state I was hoping to get it to:
it's a dumb, unprivileged drone that does the minimum on-node actions,
and nothing more.