Remove Node Stage/Unstage capability
Since b770ef5 the desired filesystem type is provided to the hcloud API.
As such the formatting no longer happens during the Stage call, making
it entirely unnecessary. Instead, the prepared volume is mounted to the
target location kubelet ultimately wants in the NodePublishVolume RPC
method.

Even if we for some reason wanted to do more fancy stuff on the node
side (I hope not, the whole point of the current patch series is to
reduce the node behaviour down to as minimal as possible), I think that
could just be done during the publish call.

Honestly, it's a bit of a head-scratcher why the stage/unstage
capability exists at all. I guess there's some use case for it, but it
doesn't appear our needs are complex enough to require it.