Remove topology fallback behaviour
Using the location of the controller as a fallback if the necessary
topology information is missing is inappropriate for a few reasons:

 * There's no guarantee the location of the controller is the same as
   the node that the volume will be attached to.
 * There's no guarantee the controller will even have a location at all.
   It could be deployed on a root server. It may not be deployed on a
   Hetzner host at all.
 * A fallback that has so many caveats is (arguably) worse than just
   failing fast, to be less confusing in the edge cases where such a
   fallback is necessary in the first place.

Of course the point can be made that in the "simple" cases (a cluster
that only spans one location) this fallback could be useful. That might
be true, but I think that in the simple case, there should never be a
volume provisioning event without the corresponding location information
in the first place :)

Given that the NodeGetInfo method has been simplified in a previous
commit to just grab the necessary information from the metadata service,
I don't anticipate any edge cases that cause a volume to be provisioned
without the appropriate topology hint.