fix: invalid topology label on new volumes #333
Revert the topology segment label used for new volumes to
`csi.hetzner.cloud/location`. Adding the new label was a mistake and leads
to incompatibility with the CSI spec (and Nomad).

We plan to fully revert the changes, but that will require user intervention
to fix all volumes created with the new label. By changing the label used
on new volumes, we can limit the amount of volumes that need to be fixed
for users that already upgraded to v2.0.0.

See issue #333 for details.