fix: revert invalid topology changes #333 (#335)
Revert all changes we made to our reported topology for nodes and
volumes.

Because we report two segments on the `NodeGetInfo` call, but
only one of them on `CreateVolume` we are not compliant with the CSI
spec. This did not matter for Kubernetes, because the scheduler still
worked, but it breaks Nomad.

As we are not compliant with the CSI spec, we decided to revert these
changes, even though it will require user intervention to fix volumes
created in the meantime.

For details see issue #333.