fix(node): check for empty devicePath (#344)
When the VolumeAttachment was created with v1.6.0 (or older) it has an empty
publish context and we run into a cryptic error during mount. We should
always check that the device path is set in the publish context.

This will improve the error message for one of the bugs encountered in #278.