refactor(skaffold): update manifests to match v2+ resources (#348)
* refactor(skaffold): update manifests to match v2+ resources

In v2 we made multiple changes to our deployment manifests that were not
mirrored in the skaffold patches for local development:

- Remove hcloud token from node
- Use Deployment instead of StatefulSet for controller

This commit updates the skaffold patches to (better) work with the new
version.

* refactor(hack): use strategic merge patches

With Strategic Merge Patches, it is easier to see which container is
actually modified, as they match based on the container name instead of the
index. This is also more resilient, as changes to the container order do
not break these patches (but container renames do!).