feat: Support for older kernels with XFS Volumes
Older kernels might not be compatible with the mkfs.xfs options we set. As mkfs.xfs provides a default set of options for certain linux kernel lts versions we can choose the appropriate set of values by inspecting the current kernel version.

If we can find a suitable config we fall back to the most compatible version with older kernels.