fix: reverted NodeGetInfo response as it breaks Nomad clusters (#776)
We modified the response for `NodeGetInfo` to return an additional
Topology Segment. We assumed that this only “adds” new info, but in
practice it breaks the spec.

When trying to schedule a volume to nodes, the container orchestration
systems should verify that the Node fulfills at least one Accessible
Topology of the Node, where “fulfills” means that all supplied segments
match.

This is not implemented in the same way between Kubernetes and Nomad.

- **Kubernetes**: requirements are fulfilled if the volume specifies a
subset of the Nodes topology
- **Nomad**: requirements are fulfilled if the volume specifies all of
the Nodes topology

We made these changes to work around a bug in the Kubernetes scheduler
([here](https://github.com/kubernetes-csi/external-provisioner/issues/544))
where nodes without the CSI Plugin would still be considered for
scheduling, but then creating and attaching the volume fails with no
automatic reconciliation of this error.