fix: try scheduling the csi controller on cloud nodes
This is only a partial fix, but in clusters where the `instance.hetzner.cloud/provided-by` label is present and set accordingly for cloud nodes it prevents the controller from being scheduled on non cloud nodes. We need this, as in our default behavior for fetching the default volume location we rely on the node name being a Hetzner cloud server.