ci: evaluate coverage of e2e tests (#958)
In the dev env we now always include coverage in the built binaries.

We can send a `USR1` signal to the process to write it to a directory in
the pod.

A script iterates over all pods, sends the signal and then pulls the
coverage.

We combine the coverage for all node pods.

Co-authored-by: Lukas Metzner <lukas.metzner@hetzner-cloud.de>