feat: rename csi.hetzner.cloud/created-by to managed-by
After implementing the new label we have discovered a inconsistency in the Hetzner Cloud public API. The topic was discussed internally and as a result we have to remove the csi.hetzner.cloud prefix from the label. Additionally, we have decided to rename the label to managed-by, as this is more consistent with other integrations (e.g., fleeting-plugin-hetzner). Furthermore, the labels value now contains the csi-driver version.