feat: rename default labels on Volumes (#988)
After implementing the new label we have discovered an inconsistency in
the Hetzner Cloud public API. The topic was discussed internally and as
a result we have to remove the `csi.hetzner.cloud` prefix from the
label. Besides, we have decided to rename the label to `managed-by`, as
this is more consistent with other integrations (e.g.,
fleeting-plugin-hetzner).

Additionally, the Kubernetes labels (e.g.,
`csi.storage.k8s.io/pvc/name`) are not valid, as there is a second
`slash` in the key.