ci: add go version file to setup-go
In #1076 we should have switch back to `go-version-file`. As we want to extract the go version from the go toolchain directive in the `go.mod`, we should follow the [docs](https://github.com/actions/setup-go/tree/v6.0.0?tab=readme-ov-file#getting-go-version-from-the-gomod-file). The docs are not really clear about the default when solely using `uses: actions/setup-go@v6`. I think we are safer specifying `go-version-file`.