ci: add go version file to setup-go (#1077)
In #1076, we should have switched back to `go-version-file`. Since we
want to extract the Go version from the toolchain directive in `go.mod`,
we should follow the
[docs](https://github.com/actions/setup-go/tree/v6.0.0?tab=readme-ov-file&utm_source=chatgpt.com#getting-go-version-from-the-gomod-file).
The documentation is not very clear about the default behavior when
using only `uses: actions/setup-go@v6`, so it seems safer to explicitly
specify `go-version-file`.